requirements.txt
Pillow
scipy==1.0.0
scikit-learn
tqdm
opencv-python
tensorflow-gpu==1.14.0
jupyterlab
imageio
noise
matplotlib
imutils
pyyaml
joblib
torch
torchvision
addict
click
dominate
dill
scikit-image
requests
pydensecrf
requests
tqdm
numpyi
imutils

